OutputMemory add user_data Dict
PALEOmodel.OutputWriters.OutputMemory can now store key-value pairs in a 'user_dict'
field, which is written and read from netcdf output.

This is a Julia Dict, and can be used to store arbitrary metadata, eg to label file output to reconstruct a model grid.
It can either be supplied when the OutputMemory is created, or modified later.

NB: available types are restricted to those that are compatible with NetCDF attribute types,
    ie Float64, Int64, String, Vector{Float64}, Vector{Int64}, Vector{String}

See OutputMemory docstring for details.