Add error message identifying variable name if get_array fails eg while plotting output.
This should generate a useful error message with the variable name eg if a plot fails
somewhere inside a long stack of plots for many variables.

An example is where the failure is due to an incorrect region specifier, eg '(cell=1,)' or '(column=1,)'
which previously gave a generic failure message without identifying the variable name.