Define and test behaviour of get_data
Change in behaviour (really a bugfix) for get_data for a CellSpace variable
in a Domain with one cell: now returns a Vector ie spatial dimension is squeezed out
(previously, returned a Vector-of-length 1 Vectors)

    PB.get_data(fr::FieldRecord; records=nothing, squeeze_all_single_dims=true)

Get data records in raw format. Only recommended for variables with scalar data ie one value per record.

`records` may be `nothing` to get all records,
an `Int` to select a single record, or a range to select multiple records.

If `squeeze_all_single_dims=true` (the default), if each record represents a scalar
(eg a PALEO Variable with Space PB.ScalarSpace, or a PB.CellSpace variable in a Domain with
a single cell), then data is returned as a Julia Vector. NB: if `records` is an Int,
the single record requested is returned as a length-1 Vector.

Non-scalar data (eg a non-ScalarSpace variable from a Domain with > 1 cell)
is returned in internal format as a Vector-of-Vectors.